Text Chat System
There are two programs in the TCS. The first program is TCS server, it authenticates the users, update the list of online users to all the currently online, create a connection between sender and receiver, store the offline messages. To reduce the server system resources, the chat server should be a text based UI. The second one is the TCS client; it provides login, view list users, and send message to a selected user.